// src/interfaces/AspError.ts
//----------------------
// <auto-generated>
//     Generated using the NSwag toolchain v10.9.0.0 (Newtonsoft.Json v13.0.0.0) (http://NJsonSchema.org)
// </auto-generated>
//----------------------







export interface AspError {
    fail: boolean;
    route: string;
    traceId: string;
    code: string;
    description: string;
}